Extract BookmarkButton into reusable component
