+++ src/components/dictionary-view/dictionary-view.tsx
@@ -1,9 +1,11 @@
 import { useState, useEffect, useContext } from 'react';
+import { useOutletContext } from 'react-router-dom';
+
+// FPCC
 import WordCardMobile from '../dictionary-page/word-card-mobile';
 import WordCardDesktop from '../dictionary-page/word-card-desktop';
 import { DictionaryType, FvWord, isFvWordLocationCombo } from '../common/data';
 import MultiSwitch from '../common/multi-switch/multi-switch';
-import { useOutletContext } from 'react-router-dom';
 import fetchWordsData from '../../services/wordsApiService';
 import { SearchContext } from '../search-provider';
 import { LoadingSpinner } from '../common/loading-spinner/loading-spinner';
